Moved Serializers assets to Assets dir
