+++ tests/Feature/Serializers/JsonSerializerTest.php
@@ -2,8 +2,8 @@
 
 namespace LaravelDoctrineTest\ORM\Feature\Serializers;
 
-use LaravelDoctrine\ORM\Serializers\Jsonable;
 use LaravelDoctrine\ORM\Serializers\JsonSerializer;
+use LaravelDoctrineTest\ORM\Assets\Serializers\JsonableEntity;
 use LaravelDoctrineTest\ORM\TestCase;
 
 class JsonSerializerTest extends TestCase
@@ -39,31 +39,5 @@ class JsonSerializerTest extends TestCase
 
         $this->assertJson($json);
         $this->assertEquals('{"id":"IDVALUE","name":"NAMEVALUE","numeric":1}', $json);
-    }
-}
-
-class JsonableEntity
-{
-    use Jsonable;
-
-    protected $id = 'IDVALUE';
-
-    protected $name = 'NAMEVALUE';
-
-    protected $numeric = "1";
-
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    public function getNumeric()
-    {
-        return $this->numeric;
     }
 }
